requirements.txt
numpy
pillow
pyqtdarktheme @ git+https://github.com/5yutan5/PyQtDarkTheme.git
pyside6
pytoshop
rich
setuptools
six
unitypy